refactor: event 패키지로 관리
